pyproject.toml
[project]
name = "rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-storage-blob==12.26.0",
    "chromadb==1.0.15",
    "docling-ocr-onnxtr>=0.2.0",
    "fastapi==0.115.8",
    "giskard[llm]>=2.18.0",
    "huggingface-hub==0.28.1",
    "langchain==0.3.17",
    "langchain-community==0.3.16",
    "langchain-core==0.3.33",
    "langchain-docling==1.0.0",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.2.14",
    "langchain-text-splitters==0.3.5",
    "langgraph==0.2.69",
    "langserve==0.3.1",
    "matplotlib==3.10.3",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "ragas>=0.3.7",
    "scikit-learn==1.7.1",
    "sentence-transformers==5.0.0",
    "sqlmodel==0.0.24",
    "sse-starlette>=3.0.2",
    "tenacity==9.1.2",
    "torch==2.7.1",
    "transformers==4.50.0",
]
